requirements.txt
asgiref==3.8.1
click==8.1.7
Django==5.0.4
djangorestframework==3.15.1
ghp-import==2.1.0
Jinja2==3.1.3
Markdown==3.6
MarkupSafe==2.1.5
mergedeep==1.3.4
mkdocs==1.5.3
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyyaml_env_tag==0.1
six==1.16.0
sqlparse==0.4.4
watchdog==4.0.0
